feat: cloudinary implemented

+++ server/controllers/categories.js
@@ -1,12 +1,13 @@
+import { deleteFileFromCloudinary, uploadImageToCloudinary } from "../helpers/functions.js";
 import Category from "../models/Category.js";
 import fs from "fs";
 export const createCategory = async (req, res) => {
   const file = req.file;
   if (!file) return res.status(400).json({ message: "No icon found" });
-  const fileName = file.filename;
+  const fileFromCloudinary = await uploadImageToCloudinary(req.file,res)
   let category = new Category({
     name: req.body.name,
-    icon: fileName,
+    icon: fileFromCloudinary?.url,
     color: req.body.color,
     description: req.body.description,
   });
@@ -14,7 +15,6 @@ export const createCategory = async (req, res) => {
 
   if (!category)
     return res.status(400).json({ message: "the category cannot be created!" });
-
   res.status(201).json(category);
 };
 export const getCategory = async (req, res) => {
@@ -39,10 +39,10 @@ export const updateCategory = async (req, res) => {
   let iconToSet;
   const file = req.file;
   if (file) {
-    const fileName = file.filename;
-    if (fileName) {
-      fs.unlinkSync("./public/uploads/" + oldCategory.icon);
-      iconToSet = fileName;
+    const isDeleted = await deleteFileFromCloudinary(oldCategory?.icon)
+    if (isDeleted){
+      const fileFromCloudinary = await uploadImageToCloudinary(req.file, res)
+      iconToSet = fileFromCloudinary?.url;
     }
   } else {
     iconToSet = req.body.icon;
@@ -59,27 +59,47 @@ export const updateCategory = async (req, res) => {
   );
 
   if (!category)
-    return res.status(400).json({ message: "the category cannot be created!" });
+    return res.status(400).json({ message: "the category cannot be updated!" });
   res.status(201).json(category);
 };
 export const deleteCategory = async (req, res) => {
-  const oldCategory = await Category.findById(req.params.id);
-  if (oldCategory) {
-    fs.unlinkSync("./public/uploads/" + oldCategory.icon);
-  }
-  Category.findByIdAndRemove(req.params.id)
-    .then((category) => {
-      if (category) {
-        return res
-          .status(200)
-          .json({ success: true, message: "The category is deleted!" });
-      } else {
-        return res
-          .status(404)
-          .json({ success: false, message: "category not found!" });
+  try {
+    const oldCategory = await Category.findById(req.params.id);
+    if (oldCategory) {
+      const isDeleted = await deleteFileFromCloudinary(oldCategory?.icon)
+      if(isDeleted){
+        Category.findByIdAndRemove(req.params.id)
+          .then((category) => {
+            if (category) {
+              return res
+                .status(200)
+                .json({ success: true, message: "The category is deleted!" });
+            } else {
+              return res
+                .status(404)
+                .json({ success: false, message: "category not found!" });
+            }
+          })
+          .catch((err) => {
+            return res.status(500).json({ success: false, error: err });
+          });
+      }else{
+        res
+          .status(500)
+          .json({ message: "Not able to delete image" });
       }
-    })
-    .catch((err) => {
-      return res.status(500).json({ success: false, error: err });
-    });
+    }else{
+      res
+        .status(500)
+        .json({ message: "Not found any category" });
+
+    }
+   
+  } catch (error) {
+      res
+      .status(500)
+      .json({ message: "Internal server error" });
+    
+  }
+ 
 };
