+++ server/controllers/subCategories.js
@@ -1,13 +1,12 @@
-import Category from "../models/Category.js";
+import { deleteFileFromCloudinary, uploadImageToCloudinary } from "../helpers/functions.js";
 import SubCategory from "../models/SubCategory.js";
-import fs from "fs";
 export const createSubCategory = async (req, res) => {
   const file = req.file;
   if (!file) return res.status(400).json({ message: "No icon found" });
-  const fileName = file.filename;
+  const fileFromCloudinary = await uploadImageToCloudinary(req.file, res)
   let category = new SubCategory({
     name: req.body.name,
-    icon: fileName,
+    icon: fileFromCloudinary?.url,
     color: req.body.color,
     category: req.body.category,
     description: req.body.description,
@@ -50,10 +49,10 @@ export const updateSubCategory = async (req, res) => {
   let iconToSet;
   const file = req.file;
   if (file) {
-    const fileName = file.filename;
-    if (fileName) {
-      fs.unlinkSync("./public/uploads/" + oldCategory.icon);
-      iconToSet = fileName;
+    const isDeleted = await deleteFileFromCloudinary(oldCategory?.icon)
+    if (isDeleted) {
+      const fileFromCloudinary = await uploadImageToCloudinary(req.file, res)
+      iconToSet = fileFromCloudinary?.url;
     }
   } else {
     iconToSet = req.body.icon;
@@ -71,27 +70,46 @@ export const updateSubCategory = async (req, res) => {
   );
 
   if (!category)
-    return res.status(400).json({ message: "the category cannot be created!" });
+    return res.status(400).json({ message: "the category cannot be updated!" });
   res.status(201).json(category);
 };
 export const deleteSubCategory = async (req, res) => {
-  const oldSubCategory = await SubCategory.findById(req.params.id);
-  if (oldSubCategory) {
-    fs.unlinkSync("./public/uploads/" + oldSubCategory.icon);
-  }
-  SubCategory.findByIdAndRemove(req.params.id)
-    .then((category) => {
-      if (category) {
-        return res
-          .status(200)
-          .json({ success: true, message: "the category is deleted!" });
+  try {
+    const oldSubCategory = await SubCategory.findById(req.params.id);
+    if (oldSubCategory) {
+      const isDeleted = await deleteFileFromCloudinary(oldSubCategory?.icon)
+      if (isDeleted) {
+        SubCategory.findByIdAndRemove(req.params.id)
+          .then((category) => {
+            if (category) {
+              return res
+                .status(200)
+                .json({ success: true, message: "The category is deleted!" });
+            } else {
+              return res
+                .status(404)
+                .json({ success: false, message: "category not found!" });
+            }
+          })
+          .catch((err) => {
+            return res.status(500).json({ success: false, error: err });
+          });
       } else {
-        return res
-          .status(404)
-          .json({ success: false, message: "category not found!" });
+        res
+          .status(500)
+          .json({ message: "Not able to delete image" });
       }
-    })
-    .catch((err) => {
-      return res.status(500).json({ success: false, error: err });
-    });
+    } else {
+      res
+        .status(500)
+        .json({ message: "Not found any category" });
+
+    }
+
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Internal server error" });
+
+  }
 };
